refactor(atom-cache): move root-resolver to services directory
Relocated root-resolver from util/ to services/ as requested during architectural review. This aligns with project standards for modules performing complex logic and I/O. Updated all internal and test references.

Lore-id: ebe7b9c4
Constraint: Root resolution must be non-throwing (fallback to CWD on error)
Tested: Successfully moved src/util/root-resolver.ts to src/services/
Tested: Updated imports in src/main.ts and tests/unit/root-resolver.test.ts
Tested: Verified all 5 root-resolver unit tests pass
Tested: Full project suite (452 tests) passing
Confidence: high
